+++ crates/core/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - `version` ([#476](https://github.com/stac-utils/stac-rs/pull/476))
+- Permissive deserialization ([#505](https://github.com/stac-utils/stac-rs/pull/505))
 
 ### Removed